add missing packages to the minimal requirements example
When running a bare-minimum debian like the stretch-slim image from
docker.hub, some requirements are missing to complete the installation.

apt-key will fail to work without a gpg implementation and the dirmngr
for tmp-dir creation installed. ca-certificates are not included, which
will make any `apt update` fail with "The repository '<...> stable
Release' does not have a Release file."

I tried to harmonize the addition with the checks already present.